sunxi: cedar: Allow to use CMA for cedar memory allocation
In the case if the kernel supports CMA and has it enabled, now
the cedar module tries to allocate the 80MB memory buffer from CMA.
This memory block is kept in use as long as the cedar module is
loaded. When the module is unloaded, the allocated memory gets
released.

It is preferable to allocate cedar memory buffer from the first
256MB of physical memory to avoid bugs even though in theory
higher addresses might be supported:
    http://irclog.whitequark.org/linux-sunxi/2013-07-17#4409315;
In practice, both VLC (using the proprietary libve library) and
libvdpau-sunxi fail for the buffers allocated from above 256MB
(or partially crossing this boundary).

For the kernels without CMA, the old behaviour is preserved.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>