sunxi: Unbreak broken CedarX userland software (2GB phys->bus issue)
The commit 1e205f75658cb0cf3addc28ddaa9912ca9ce14e8 happened to
break broken CedarX media players on the systems with 2GB of RAM
(Cubietruck). The problem is that some CedarX userland code (used
in VLC and XBMC together with proprietary blobs) already feeds
bus addresses to the sunxi disp ioctls, while physical addresses
are expected there. The open source libvdpau-sunxi does not have
such bugs.

This patch avoids phys->bus conversion if the address passed to
the __phys_to_bus macro is already obviously a bus address.

While the right solution would be to hunt down and fix all the
userland offenders, this may be difficult in practice. For example
when dealing with Android images and trying to provide a drop-in
kernel replacement for them.

Signed-off-by: Siarhei Siamashka <siarhei.siamas...@gmail.com>